requirements.txt
autocorrect==0.4.4
Flask==1.1.1
Flask-RESTful==0.3.8
gunicorn==20.0.4
opencv-contrib-python-headless==4.2.0.32
Pillow==7.0.0
pytesseract==0.3.2
pyenchant==3.0.1
